+++ src/main/java/org/black_ixx/bossshop/managers/misc/StringManager.java
@@ -10,6 +10,7 @@ import org.black_ixx.bossshop.events.BSCheckStringForFeaturesEvent;
 import org.black_ixx.bossshop.events.BSTransformStringEvent;
 import org.black_ixx.bossshop.managers.ClassManager;
 import org.black_ixx.bossshop.managers.serverpinging.ConnectedBuyItem;
+import org.black_ixx.bossshop.misc.ClassAndVerTools;
 import org.black_ixx.bossshop.misc.MathTools;
 import org.black_ixx.bossshop.misc.Misc;
 import org.bukkit.Bukkit;
@@ -75,10 +76,12 @@ public class StringManager {
     }
 
     private String colorize(String string) {
-        Pattern pattern = Pattern.compile("#[a-fA-F0-9]{6}");
-        for (Matcher matcher = pattern.matcher(string); matcher.find(); matcher = pattern.matcher(string)) {
-            String color = string.substring(matcher.start(), matcher.end());
-            string = string.replace(color, net.md_5.bungee.api.ChatColor.of(color)+"");
+        if (ClassAndVerTools.isHighThanOrEquals(16)) {
+            Pattern pattern = Pattern.compile("#[a-fA-F0-9]{6}");
+            for (Matcher matcher = pattern.matcher(string); matcher.find(); matcher = pattern.matcher(string)) {
+                String color = string.substring(matcher.start(), matcher.end());
+                string = string.replace(color, net.md_5.bungee.api.ChatColor.of(color) + "");
+            }
         }
         string = ChatColor.translateAlternateColorCodes('&', string);
         return string;
