+++ src/main/java/org/black_ixx/bossshop/misc/ClassAndVerTools.java
@@ -1,0 +1,21 @@
+package org.black_ixx.bossshop.misc;
+
+import org.bukkit.Bukkit;
+
+public class ClassAndVerTools {
+    public static boolean classExists(String className) {
+        try {
+            Class.forName(className);
+            return true;
+        } catch (ClassNotFoundException e) {
+            return false;
+        }
+    }
+
+    public static boolean isHighThanOrEquals(int i){
+        String version = Bukkit.getServer().getBukkitVersion().split("-")[0];
+        int version2 = Integer.parseInt(version.split("\\.")[1]);
+        return version2 >= i;
+    }
+
+}
